+++ extensions/browser-bookmarks/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Browser Bookmarks Changelog
 
-## [Support for Arc] - {PR_MERGE_DATE}
+## [Improvements] - 2023-10-18
+
+- Always sort the bookmarks by their frecency
+
+## [Support for Arc] - 2023-09-22
 
 - Add support for `Arc` browser